重构(playground): 将 Embedding 配置面板提取为独立组件
- 将配置面板逻辑从 embedding 页面中分离
- 创建 EmbeddingConfigPanel 组件，提高代码复用性
- 优化代码结构，增强可维护性
